move fts for parser and added matrix parser
