+++ Makefile
@@ -12,9 +12,11 @@ MLIBX_DIR = ./minilibx-linux
 
 SRCS = $(SRCS_DIR)/main.c \
 	$(SRCS_DIR)/ft_handle_error.c \
-	$(SRCS_DIR)/ft_loadmap.c \
-	$(SRCS_DIR)/ft_loadmap_utils.c \
-	$(SRCS_DIR)/toolbox.c \
+	$(SRCS_DIR)/parser/ft_loadmap.c \
+	$(SRCS_DIR)/parser/ft_map_parser.c \
+	$(SRCS_DIR)/parser/ft_matrix_parser.c \
+	$(SRCS_DIR)/parser/ft_matrix_parser_utils.c \
+	$(SRCS_DIR)/parser/parser_utils.c \
 
 #B_SRCS = $(BONUS_DIR)/main_bonus.c \
