+++ srcs/main.c
@@ -1,4 +1,4 @@
-#include "../includes/cub3D.h"
+#include "../includes/cub3d.h"
 
 int	main(int argc, char **argv)
 {
